requirements.txt
argcomplete==1.12.0
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.1
cycler==0.10.0
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.5
django-bootstrap4==2.2.0
django-cleanup==5.0.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-heroku==0.3.1
django-mongoengine==0.4.2
django-precise-bbcode==1.2.13
django-ranged-response==0.2.0
django-rest-framework-mongoengine==3.4.1
django-simple-captcha==0.5.12
djangorestframework==3.11.1
djongo==1.3.3
dnspython==2.0.0
easy-thumbnails==2.7
gunicorn==20.0.4
heroku==0.1.4
httpie==2.2.0
idna==2.10
importlib-metadata==1.7.0
isort==5.5.2
joblib==0.16.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mongoengine==0.20.0
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
pipx==0.15.5.1
psycopg2==2.8.6
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
threadpoolctl==2.1.0
toml==0.10.1
transliterate==1.10.2
typed-ast==1.4.1
urllib3==1.25.10
userpath==1.4.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.2.0
